+++ sei-ibc-go/CHANGELOG.md
@@ -38,6 +38,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Dependencies
 
+* [\#485](https://github.com/cosmos/ibc-go/pull/485) Bump SDK version to v0.44.2
 * [\#455](https://github.com/cosmos/ibc-go/pull/455) Bump SDK version to v0.44.1
 
 ### API Breaking