+++ sei-ibc-go/CHANGELOG.md
@@ -38,7 +38,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Dependencies
 
-* [\#485](https://github.com/cosmos/ibc-go/pull/485) Bump SDK version to v0.44.2
+* [\#503](https://github.com/cosmos/ibc-go/pull/503) Bump SDK version to v0.44.3
 * [\#455](https://github.com/cosmos/ibc-go/pull/455) Bump SDK version to v0.44.1
 * [\#404](https://github.com/cosmos/ibc-go/pull/404) Bump Go version to 1.17